de-cake environments
